Starting unit testing
Basic testing is derived from the Laracast's testing tutorials, in
addition to the basic validations code, however, there are some tweaks.

* If you pass an array to the ApiController::respondWithErrors method,
it'll pass each one through the translation engine.
* I'm not exactly sure why we're doing Talk::checkFormatting - I think
it should do this automatically, but just to be on the safe side, leave
this in :)
* In TalkTest::it_creates_a_new_talk_given_valid_parameters - check it
gets the ID back, and the attributes we're expecting.

I've added some new response messages (and fixed the 201 code, spotted
by the unit test!), and I've unmangled the fields we return (we don't
need that pesky backwards compatibility... right?) in both the
Transformers\talkTransformer and TalkTest::getStub()